refactor(components/molecule/inputTags): add import file extensions
